requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=1.0.0
langchain-mcp-adapters>=0.1.11
langgraph>=0.0.26
mcp>=0.9.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
cryptography>=41.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pydantic>=2.0.0
chromadb>=0.4.0
